+++ src/main/java/subway/controller/StationController.java
@@ -1,9 +1,13 @@
-package subway.station;
+package subway.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import subway.domain.Station;
+import subway.dao.StationDao;
+import subway.request.StationRequest;
+import subway.request.StationResponse;
 
 import java.net.URI;
 import java.util.List;
