+++ src/main/java/subway/dao/LineDao.java
@@ -1,4 +1,4 @@
-package subway.line;
+package subway.dao;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.EmptyResultDataAccessException;
@@ -7,6 +7,7 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
+import subway.domain.Line;
 
 import java.sql.PreparedStatement;
 import java.sql.Statement;
