+++ src/main/java/subway/dao/SectionDao.java
@@ -1,12 +1,12 @@
-package subway.section;
+package subway.dao;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
+import subway.domain.Section;
 
 import java.sql.PreparedStatement;
 import java.sql.Statement;
