+++ src/main/java/subway/domain/Station.java
@@ -1,4 +1,4 @@
-package subway.station;
+package subway.domain;
 
 public class Station {
     private Long id;
