+++ src/main/java/subway/request/LineRequest.java
@@ -1,4 +1,4 @@
-package subway.line;
+package subway.request;
 
 public class LineRequest {
     private String name;
