+++ src/main/java/subway/request/StationRequest.java
@@ -1,4 +1,4 @@
-package subway.station;
+package subway.request;
 
 public class StationRequest {
     private String name;
