+++ src/main/java/subway/service/SectionService.java
@@ -1,8 +1,10 @@
-package subway.section;
+package subway.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import subway.dao.SectionDao;
+import subway.domain.Section;
 
 import java.util.ArrayList;
 import java.util.HashMap;
