+++ src/test/java/subway/line/LineAcceptanceTest.java
@@ -9,7 +9,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import subway.AcceptanceTest;
-import subway.station.StationResponse;
+import subway.request.LineRequest;
+import subway.response.LineResponse;
+import subway.request.StationResponse;
 
 import java.util.Arrays;
 import java.util.List;
